satellite/sound-presence/src/requirements_all.txt
numpy
matplotlib
scipy
sklearn
hmmlearn
simplejson
eyed3
pydub
pyalsaaudio
paho-mqtt
pyyaml
